LayoutTests/js/typedarray-filter-expected.txt

This test checks the behavior of the TypedArray.prototype.filter function

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Int32Array.prototype.filter.length is 1
PASS Int32Array.prototype.filter.name is 'filter'
PASS isSameFunctionForEachTypedArrayPrototype('filter') is true
PASS testPrototypeReceivesArray('filter', [undefined, this, { }, [ ], true, '']) is true

testPrototypeFunction has the following arg list (name, args, init, result [ , expectedArray ])

1.0 Single Argument Testing
PASS testPrototypeFunction('filter', '(keepEven)', [12, 5, 8, 13, 44], [12, 8, 44]) is true
PASS testPrototypeFunction('filter', '(keepEven)', [11, 54, 18, 13, 1], [54, 18]) is true

2.0 Two Argument Testing
PASS testPrototypeFunction('filter', '(keepEven, thisValue)', [12, 23, 11, 1, 45], [12, 23, 1]) is true

3.0 Array Element Changing
PASS testPrototypeFunction('filter', '(keepEvenAndChange)', [12, 15, 2, 13, 44], [12, 2], [5, 5, 5, 5, 5]) is true

4.0 Exception Test
PASS testPrototypeFunction('filter', '(isBigEnoughAndException)', [12, 15, 10, 13, 44], false) threw exception exception from function.

5.0 Wrong Type for Callback Test
PASS testPrototypeFunction('filter', '(8)', [12, 15, 10, 13, 44], false) threw exception TypeError: Array.prototype.map callback must be a function.
PASS testPrototypeFunction('filter', '("wrong")', [12, 15, 10, 13, 44], false) threw exception TypeError: Array.prototype.map callback must be a function.
PASS testPrototypeFunction('filter', '(new Object())', [12, 15, 10, 13, 44], false) threw exception TypeError: Array.prototype.map callback must be a function.
PASS testPrototypeFunction('filter', '(null)', [12, 15, 10, 13, 44], false) threw exception TypeError: Array.prototype.map callback must be a function.
PASS testPrototypeFunction('filter', '()', [12, 15, 10, 13, 44], false) threw exception TypeError: Array.prototype.map callback must be a function.

PASS successfullyParsed is true

TEST COMPLETE
